Add atomic locks around folder cleanup.

+++ gfx.cpp
@@ -9880,19 +9880,24 @@ private:
                 fclose(fd); // write out PDB for shader debugging
             }
             // Delete older pdb files
-            std::filesystem::directory_iterator const end;
-            std::vector<std::filesystem::path>        pdb_files;
-            std::error_code                           ec, ec2;
-            for(std::filesystem::directory_iterator iter{std::filesystem::path(shader_pdb_dir + shader_key_dir)}; iter != end && !ec2; iter.increment(ec2))
-                if(is_regular_file(*iter, ec))
-                    if(iter->path().extension() == ".pdb")
-                        pdb_files.emplace_back(*iter);
-            if(pdb_files.size() > max_cached_files)
+            static std::atomic_bool shader_pdb_cleanup;
+            if(bool expected = false; !shader_pdb_cleanup.compare_exchange_strong(expected, true))
             {
-                std::sort(pdb_files.begin(), pdb_files.end(), [&ec](const std::filesystem::path &a, const std::filesystem::path &b) {
-                    return last_write_time(a, ec) > last_write_time(b, ec); });
-                for(size_t i = max_cached_files; i < pdb_files.size(); ++i)
-                    std::filesystem::remove(pdb_files[i], ec);
+                std::filesystem::directory_iterator const end;
+                std::vector<std::filesystem::path>        pdb_files;
+                std::error_code                           ec, ec2;
+                for(std::filesystem::directory_iterator iter{std::filesystem::path(shader_pdb_dir + shader_key_dir)}; iter != end && !ec2; iter.increment(ec2))
+                    if(is_regular_file(*iter, ec))
+                        if(iter->path().extension() == ".pdb")
+                            pdb_files.emplace_back(*iter);
+                if(pdb_files.size() > max_cached_files)
+                {
+                    std::sort(pdb_files.begin(), pdb_files.end(), [&ec](const std::filesystem::path &a, const std::filesystem::path &b) {
+                        return last_write_time(a, ec) > last_write_time(b, ec); });
+                    for(size_t i = max_cached_files; i < pdb_files.size(); ++i)
+                        std::filesystem::remove(pdb_files[i], ec);
+                }
+                shader_pdb_cleanup.exchange(false);
             }
         }
 
@@ -9924,32 +9929,37 @@ private:
                         fclose(fd); // write out reflection for shader caching
                     }
                     // Delete older cached files
-                    std::filesystem::directory_iterator const end;
-                    std::vector<std::filesystem::path>        bytecode_files;
-                    std::vector<std::filesystem::path>        reflection_files;
-                    std::error_code                           ec, ec2;
-                    for(std::filesystem::directory_iterator iter{std::filesystem::path(shader_cache_dir + shader_key_dir)}; iter != end && !ec2; iter.increment(ec2))
-                        if(is_regular_file(*iter, ec))
+                    static std::atomic_bool shader_cache_cleanup;
+                    if(bool expected = false; !shader_cache_cleanup.compare_exchange_strong(expected, true))
+                    {
+                        std::filesystem::directory_iterator const end;
+                        std::vector<std::filesystem::path>        bytecode_files;
+                        std::vector<std::filesystem::path>        reflection_files;
+                        std::error_code                           ec, ec2;
+                        for(std::filesystem::directory_iterator iter{std::filesystem::path(shader_cache_dir + shader_key_dir)}; iter != end && !ec2; iter.increment(ec2))
+                            if(is_regular_file(*iter, ec))
+                            {
+                                if(iter->path().extension() == ".bytecode")
+                                    bytecode_files.emplace_back(*iter);
+                                else if(iter->path().extension() == ".reflection")
+                                    reflection_files.emplace_back(*iter);
+                            }
+                        if(bytecode_files.size() > max_cached_files)
                         {
-                            if(iter->path().extension() == ".bytecode")
-                                bytecode_files.emplace_back(*iter);
-                            else if(iter->path().extension() == ".reflection")
-                                reflection_files.emplace_back(*iter);
-                        }
-                    if(bytecode_files.size() > max_cached_files)
-                    {
-                        std::sort(bytecode_files.begin(), bytecode_files.end(), [&ec](const std::filesystem::path &a, const std::filesystem::path &b) {
-                            return last_write_time(a, ec) > last_write_time(b, ec); });
-                        for(size_t i = max_cached_files; i < bytecode_files.size(); ++i)
-                            std::filesystem::remove(bytecode_files[i], ec);
-                    }
-                    if (reflection_files.size() > max_cached_files)
-                    {
-                        std::sort(reflection_files.begin(), reflection_files.end(),
-                            [&ec](std::filesystem::path const &a, std::filesystem::path const &b) {
+                            std::sort(bytecode_files.begin(), bytecode_files.end(), [&ec](const std::filesystem::path &a, const std::filesystem::path &b) {
                                 return last_write_time(a, ec) > last_write_time(b, ec); });
-                        for(size_t i = max_cached_files; i < reflection_files.size(); ++i)
-                            std::filesystem::remove(reflection_files[i], ec);
+                            for(size_t i = max_cached_files; i < bytecode_files.size(); ++i)
+                                std::filesystem::remove(bytecode_files[i], ec);
+                        }
+                        if (reflection_files.size() > max_cached_files)
+                        {
+                            std::sort(reflection_files.begin(), reflection_files.end(),
+                                [&ec](std::filesystem::path const &a, std::filesystem::path const &b) {
+                                    return last_write_time(a, ec) > last_write_time(b, ec); });
+                            for(size_t i = max_cached_files; i < reflection_files.size(); ++i)
+                                std::filesystem::remove(reflection_files[i], ec);
+                        }
+                        shader_cache_cleanup.exchange(false);
                     }
                 }
             }
